ch-#64: move useTab to hooks folder and  did cleanup
