+++ src/components/Tabs/Tabs.tsx
@@ -1,31 +1,13 @@
-import { useNavigate } from "react-router-dom";
-import { useState } from "react";
-import TabsWrapper from "../components/Tabs/TabsStyle";
+import TabsWrapper from "./TabsStyle";
+import useTabs from "../../hooks/useTabs";
 
-type TProps = {
+export type TProps = {
   data: Array<{
     id: string;
     heading: string;
   }>;
   children: Array<JSX.Element>;
   type2?: boolean;
-};
-
-const useTabs = (data: TProps["data"]) => {
-  const params = new URLSearchParams(document.location.search);
-  const navigate = useNavigate();
-  const [activeId, setActiveId] = useState(params.get("tabId") || data[0].id);
-
-  const handleChangeActiveId = (tabId: string) => {
-    setActiveId(tabId);
-    navigate({ search: `?tabId=${tabId}` });
-  };
-
-  const getActiveIndex = () => {
-    return data.findIndex((tab) => tab.id === activeId);
-  };
-
-  return { activeId, handleChangeActiveId, getActiveIndex };
 };
 
 const Tabs = ({ data, children, type2 }: TProps) => {
